edit crypto page and move between pages logic code
