+++ lib/my_app.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:flutterestudo/pages/crypto_page.dart';
 import 'package:flutterestudo/pages/login_page.dart';
 import 'package:flutterestudo/pages/main_page.dart';
 
@@ -8,6 +9,7 @@ class MyApp extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
+      debugShowCheckedModeBanner: false,
       theme: ThemeData(primarySwatch: Colors.red),
       home: const MainPage(),
     );
