+++ lib/pages/crypto_page.dart
@@ -15,6 +15,8 @@ class _CryptoPageState extends State<CryptoPage> {
   bool _isLoading = true;
   String? _errorMessage;
   Map<String, double>? _cryptoPrices;
+  List<String> priceList = [];
+  List<String> cryptoList = [];
 
   @override
   void initState() {
@@ -45,6 +47,12 @@ class _CryptoPageState extends State<CryptoPage> {
         setState(() {
           _cryptoPrices = prices;
           _isLoading = false;
+          priceList = _cryptoPrices!.entries.map((ett) {
+            return ett.value.toStringAsFixed(2);
+          }).toList();
+          cryptoList = _cryptoPrices!.entries.map((ett) {
+            return ett.key.toUpperCase();
+          }).toList();
         });
       } else {
         throw Exception(
@@ -80,16 +88,21 @@ class _CryptoPageState extends State<CryptoPage> {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
+      backgroundColor: Colors.white,
       appBar: AppBar(
-        iconTheme: const IconThemeData(color: Colors.white),
-        title: const Text(
+        iconTheme: const IconThemeData(color: Colors.black87),
+        backgroundColor: Colors.white,
+        surfaceTintColor: Color(0xffffffff),
+        title: Text(
           "CryptoPage",
-          style: TextStyle(color: Colors.white),
+          style: TextStyle(fontFamily: "Poppins", color: Colors.amber),
         ),
-        backgroundColor: Colors.black,
       ),
       body: _isLoading
-          ? const Center(child: CircularProgressIndicator())
+          ? const Center(
+              child: CircularProgressIndicator(
+              color: Colors.amber,
+            ))
           : _errorMessage != null
               ? Center(child: Text('Erro: $_errorMessage'))
               : Padding(
@@ -97,70 +110,78 @@ class _CryptoPageState extends State<CryptoPage> {
                   child: Column(
                     crossAxisAlignment: CrossAxisAlignment.start,
                     children: [
-                      const Text(
+                      Text(
                         "Cotação dos principais cryptos",
-                        style: TextStyle(
-                            fontSize: 24, fontWeight: FontWeight.bold),
+                        style: TextStyle(fontSize: 22, fontFamily: "Poppins"),
                       ),
-                      const SizedBox(height: 16),
-                      Table(
-                        border: TableBorder.all(),
-                        columnWidths: const {
-                          0: FlexColumnWidth(2),
-                          1: FlexColumnWidth(3),
-                        },
-                        children: [
-                          const TableRow(
-                            children: [
-                              Padding(
-                                padding: EdgeInsets.all(8.0),
-                                child: Text('Criptomoeda',
-                                    style:
-                                        TextStyle(fontWeight: FontWeight.bold)),
+                      SizedBox(height: 16),
+                      Container(
+                        decoration: BoxDecoration(
+                          border: Border.all(
+                              width: 2.3, color: Colors.amber.shade400),
+                        ),
+                        child: Column(
+                          children: [
+                            Padding(
+                              padding: const EdgeInsets.all(8.0),
+                              child: Row(
+                                mainAxisAlignment:
+                                    MainAxisAlignment.spaceBetween,
+                                children: [
+                                  Text('Criptomoeda',
+                                      style: TextStyle(
+                                        fontFamily: "Poppins",
+                                      )),
+                                  Text(
+                                    'Valor (\$)',
+                                    style: TextStyle(
+                                      fontFamily: "Poppins",
+                                    ),
+                                  ),
+                                ],
                               ),
-                              Padding(
-                                padding: EdgeInsets.all(8.0),
-                                child: Text('Valor (R\$)',
-                                    style:
-                                        TextStyle(fontWeight: FontWeight.bold)),
+                            ),
+                            Divider(
+                              color: Colors.amber.shade400,
+                              thickness: 2.3,
+                            ),
+                            SizedBox(
+                              height: 300,
+                              child: SingleChildScrollView(
+                                physics: BouncingScrollPhysics(),
+                                child: Column(
+                                  children:
+                                      List.generate(priceList.length, (index) {
+                                    return Padding(
+                                      padding: const EdgeInsets.symmetric(
+                                          vertical: 8.0, horizontal: 14),
+                                      child: Row(
+                                        mainAxisAlignment:
+                                            MainAxisAlignment.spaceBetween,
+                                        children: [
+                                          Text(
+                                            cryptoList[index],
+                                            style: TextStyle(
+                                                fontWeight: FontWeight.w500),
+                                          ),
+                                          Text(
+                                            priceList[index],
+                                            style: TextStyle(
+                                                fontWeight: FontWeight.w500),
+                                          ),
+                                        ],
+                                      ),
+                                    );
+                                  }),
+                                ),
                               ),
-                            ],
-                          ),
-                          ..._cryptoPrices!.entries.map((entry) {
-                            return TableRow(
-                              children: [
-                                Padding(
-                                  padding: const EdgeInsets.all(8.0),
-                                  child: Text(entry.key.toUpperCase()),
-                                ),
-                                Padding(
-                                  padding: const EdgeInsets.all(8.0),
-                                  child: Text(entry.value.toStringAsFixed(2)),
-                                ),
-                              ],
-                            );
-                          }).toList(),
-                        ],
+                            ),
+                          ],
+                        ),
                       ),
                     ],
                   ),
                 ),
-      bottomNavigationBar: BottomNavigationBar(
-        unselectedItemColor: Colors.white,
-        backgroundColor: Colors.black,
-        currentIndex: _selectedIndex,
-        onTap: _onItemTapped,
-        items: const [
-          BottomNavigationBarItem(
-            icon: Icon(Icons.money),
-            label: 'Moeda',
-          ),
-          BottomNavigationBarItem(
-            icon: Icon(Icons.currency_bitcoin),
-            label: 'Crypto',
-          ),
-        ],
-      ),
     );
   }
 }
