+++ lib/pages/main_page.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:flutterestudo/pages/coins_page.dart';
 import 'package:flutterestudo/pages/sobre_page.dart';
 import 'package:flutterestudo/pages/termos_de_uso_page.dart';
 import 'dart:convert';
@@ -13,281 +14,59 @@ class MainPage extends StatefulWidget {
   State<MainPage> createState() => _MainPageState();
 }
 
-class _MainPageState extends State<MainPage> {
-  final String _apiKey = 'a2042de51d24aec0adaebee2';
-  final String _baseUrl = 'https://v6.exchangerate-api.com/v6';
+class _MainPageState extends State<MainPage>
+    with SingleTickerProviderStateMixin {
+  List<Widget> pages = [CoinsPage(), CryptoPage()];
 
-  final List<String> _moedas = [
-    'USD',
-    'EUR',
-    'ARS',
-    'VES',
-    'GBP',
-    'JPY',
-    'CNY',
-    'CFH',
-    'CAD',
-    'ZAR',
-    'AUD'
+  int selectedIndex = 0;
+  late TabController tabController;
+  final List<IconData> icons = [
+    Icons.attach_money_sharp,
+    Icons.currency_bitcoin,
   ];
+  Widget biuldIcon(IconData icon, bool isSelct) {
+    return Icon(
+      icon,
+      size: 30,
+      color: isSelct ? Colors.amber : Colors.black,
+    );
+  }
 
-  Map<String, double>? _cotacoes;
-  bool _isLoading = true;
-  String? _errorMessage;
-  List<String> priceList = [];
-
-  int _selectedIndex = 0;
   List<NavigationDestination> pagesMenu() {
-    return [
-      NavigationDestination(
-        icon: Icon(
-          Icons.attach_money_sharp,
-          size: 30,
-          color: Colors.amber,
-        ),
-        label: 'Coins',
-      ),
-      NavigationDestination(
-        icon: Icon(Icons.currency_bitcoin),
-        label: 'Crypto',
-      )
-    ];
-  } // Indica qual aba está selecionada
+    return List.generate(
+      pages.length,
+      (index) => NavigationDestination(
+          icon: biuldIcon(icons[index], selectedIndex == index),
+          label: selectedIndex == 0 ? 'Coins' : 'Crypto'),
+    );
+  }
 
   @override
   void initState() {
     super.initState();
-    _fetchCotacoes();
-  }
 
-  Future<void> _fetchCotacoes() async {
-    try {
-      final response =
-          await http.get(Uri.parse('$_baseUrl/$_apiKey/latest/BRL'));
-
-      if (response.statusCode == 200) {
-        final data = json.decode(response.body);
-
-        if (data != null && data['conversion_rates'] is Map<String, dynamic>) {
-          final rates = data['conversion_rates'] as Map<String, dynamic>;
-
-          final Map<String, double> cotacoes = {};
-          for (String moeda in _moedas) {
-            if (rates.containsKey(moeda)) {
-              // Calcula quantos reais equivalem a 1 unidade da moeda
-              cotacoes[moeda] = 1 / rates[moeda].toDouble();
-            }
-          }
-
-          setState(() {
-            _cotacoes = cotacoes;
-            _isLoading = false;
-            priceList = _cotacoes!.entries.map((ett) {
-              return ett.value.toStringAsFixed(2);
-            }).toList();
-
-            // Atualize o estado corretamente
-          });
-        } else {
-          throw Exception('Dados inválidos recebidos da API.');
-        }
-      } else {
-        throw Exception(
-            'Erro ao buscar dados da API. Código: ${response.statusCode}');
-      }
-    } catch (e) {
-      setState(() {
-        _errorMessage = e.toString();
-        _isLoading = false; // Garanta que o estado será atualizado
-      });
-    }
-  }
-
-  void _onItemTapped(int index) {
-    if (index != _selectedIndex) {
-      // Pegando o `idUsuario` atual da rota
-      final int? idUsuario = ModalRoute.of(context)?.settings.arguments as int?;
-      if (idUsuario != null) {
-        // Navegando para a página correta e passando o `idUsuario`
-        Navigator.pushReplacement(
-          context,
-          MaterialPageRoute(
-            builder: (context) => index == 0 ? MainPage() : CryptoPage(),
-            settings: RouteSettings(arguments: idUsuario),
-          ),
-        );
-      } else {
-        print('Erro: idUsuario não encontrado');
-      }
-    }
+    tabController = TabController(length: pages.length, vsync: this);
   }
 
   @override
   Widget build(BuildContext context) {
-    final int? idUsuario = ModalRoute.of(context)?.settings.arguments as int?;
-    return SafeArea(
-        child: Scaffold(
-            backgroundColor: Color(0xffffffff),
-            drawer: Drawer(
-              child: Padding(
-                padding: EdgeInsets.symmetric(vertical: 20, horizontal: 10),
-                child: Column(
-                  crossAxisAlignment: CrossAxisAlignment.start,
-                  children: [
-                    InkWell(
-                      child: Container(
-                          padding: EdgeInsets.symmetric(vertical: 5),
-                          width: double.infinity,
-                          child: Text("Dados cadastrais")),
-                      onTap: () {
-                        Navigator.pop(context);
-                        Navigator.push(
-                            context,
-                            MaterialPageRoute(
-                                builder: (context) => DadosCadastroPage(),
-                                settings: RouteSettings(arguments: idUsuario)));
-                      },
-                    ),
-                    Divider(),
-                    InkWell(
-                      child: Container(
-                          padding: EdgeInsets.symmetric(vertical: 5),
-                          width: double.infinity,
-                          child: Text("Configurações")),
-                      onTap: () {},
-                    ),
-                    Divider(),
-                    InkWell(
-                      child: Container(
-                          padding: EdgeInsets.symmetric(vertical: 5),
-                          width: double.infinity,
-                          child: Text("Termos de uso e privacidade")),
-                      onTap: () {
-                        Navigator.pop(context);
-                        Navigator.push(
-                          context,
-                          MaterialPageRoute(
-                              builder: (context) => const TermosDeUsoPage()),
-                        );
-                      },
-                    ),
-                    Divider(),
-                    InkWell(
-                      child: Container(
-                          padding: EdgeInsets.symmetric(vertical: 5),
-                          width: double.infinity,
-                          child: Text("Sobre")),
-                      onTap: () {
-                        Navigator.pop(context);
-                        Navigator.push(
-                          context,
-                          MaterialPageRoute(
-                              builder: (context) => const sobrePage()),
-                        );
-                      },
-                    ),
-                  ],
-                ),
-              ),
-            ),
-            appBar: AppBar(
-              iconTheme: IconThemeData(color: Colors.black87),
-              backgroundColor: Colors.white,
-              surfaceTintColor: Color(0xffffffff),
-              title: Text(
-                "PAGINA PRINCIPAL",
-                style: TextStyle(fontFamily: "Poppins", color: Colors.amber),
-              ),
-            ),
-            body: _isLoading
-                ? Center(child: CircularProgressIndicator())
-                : _errorMessage != null
-                    ? Center(child: Text('Erro: $_errorMessage'))
-                    : Padding(
-                        padding: const EdgeInsets.all(16.0),
-                        child: Column(
-                          crossAxisAlignment: CrossAxisAlignment.start,
-                          children: [
-                            Text(
-                              "Cotação das principais moedas",
-                              style: TextStyle(
-                                  fontSize: 22, fontFamily: "Poppins"),
-                            ),
-                            SizedBox(height: 16),
-                            Container(
-                              decoration: BoxDecoration(
-                                border: Border.all(
-                                    width: 2.3, color: Colors.amber.shade400),
-                              ),
-                              child: Column(
-                                children: [
-                                  Padding(
-                                    padding: const EdgeInsets.all(8.0),
-                                    child: Row(
-                                      mainAxisAlignment:
-                                          MainAxisAlignment.spaceBetween,
-                                      children: [
-                                        Text('Moeda',
-                                            style: TextStyle(
-                                              fontFamily: "Poppins",
-                                            )),
-                                        Text(
-                                          'Valor (\$)',
-                                          style: TextStyle(
-                                            fontFamily: "Poppins",
-                                          ),
-                                        ),
-                                      ],
-                                    ),
-                                  ),
-                                  Divider(
-                                    color: Colors.amber.shade400,
-                                    thickness: 2.3,
-                                  ),
-                                  SizedBox(
-                                    height: 300,
-                                    child: SingleChildScrollView(
-                                      physics: BouncingScrollPhysics(),
-                                      child: Column(
-                                        children: List.generate(
-                                            priceList.length, (index) {
-                                          return Padding(
-                                            padding: const EdgeInsets.symmetric(
-                                                vertical: 8.0, horizontal: 14),
-                                            child: Row(
-                                              mainAxisAlignment:
-                                                  MainAxisAlignment
-                                                      .spaceBetween,
-                                              children: [
-                                                Text(
-                                                  _moedas[index],
-                                                  style: TextStyle(
-                                                      fontWeight:
-                                                          FontWeight.w500),
-                                                ),
-                                                Text(
-                                                  priceList[index],
-                                                  style: TextStyle(
-                                                      fontWeight:
-                                                          FontWeight.w500),
-                                                ),
-                                              ],
-                                            ),
-                                          );
-                                        }),
-                                      ),
-                                    ),
-                                  ),
-                                ],
-                              ),
-                            ),
-                          ],
-                        ),
-                      ),
-            bottomNavigationBar: NavigationBar(
-              indicatorColor: Color(0xffffffff),
-              backgroundColor: Color(0xffffffff),
-              destinations: pagesMenu(),
-            )));
+    return Scaffold(
+        backgroundColor: Color(0xffffffff),
+        body: TabBarView(
+            physics: const NeverScrollableScrollPhysics(),
+            controller: tabController,
+            children: pages),
+        bottomNavigationBar: NavigationBar(
+            onDestinationSelected: (value) {
+              setState(() {
+                selectedIndex = value;
+                tabController.index = value;
+              });
+            },
+            selectedIndex: selectedIndex,
+            height: 65,
+            indicatorColor: const Color(0xffffffff),
+            backgroundColor: const Color(0xffffffff),
+            destinations: pagesMenu()));
   }
 }
